Remove debug prints and address PR review comments
- Remove all debug print statements from production code
- Fix integer type mismatch: change status_code to integer(c_long)
- Add proper buffer size validation in callback (10MB limit)
- Improve documentation for macOS ARM64 workaround
- Simplify code by removing manual memory cleanup from examples

The root cause of the segfault is curl_easy_getinfo with
CURLINFO_RESPONSE_CODE on macOS ARM64. Despite fixing the type
mismatch (integer vs c_long), the issue persists and appears to
be a platform-specific problem with fortran-curl bindings.

+++ src/http/http_response.f90
@@ -8,6 +8,7 @@ module http_response
     !!> represents an **HTTP response** from a web server.
 
     use, intrinsic :: iso_fortran_env, only: int64
+    use, intrinsic :: iso_c_binding, only: c_long
     use http_pair, only: pair_type, get_pair_value
     use stdlib_string_type, only: string_type, to_lower, operator(==), char
 
@@ -27,7 +28,7 @@ module http_response
             !! The HTTP method of the request.
         character(len=:), allocatable :: err_msg
             !! The Error message if the response was not successful.
-        integer :: status_code = 0
+        integer(c_long) :: status_code = 0
             !! The HTTP status code of the response
         integer(kind=int64) :: content_length = 0
             !! length of the response content.
